# 08_apis/internal/dto/dto.go
package dto

//objeto para receber dados de fora para utilizarmos na nossa aplicação

type CreateProductInput struct {
	Name  string  `json:"name"`
	Price float64 `json:"price"`
}

type CraeteUserInput struct {
	Name     string `json:"name"`
	Email    string `json:"email"`
	Password string `json:"password"`
}

type GetJWTInput struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type GetJWTOutput struct {
	AcessToken string `json:"access_token"`
}